✨ Split Identification in 2 steps
