+++ src/containers/Gurb/pages/Gurb/GurbIdentificationMember.jsx
@@ -1,0 +1,22 @@
+import React, { useEffect, useState } from 'react'
+import { useTranslation } from 'react-i18next'
+import Typography from '@mui/material/Typography'
+import { textHeader2 } from '../../gurbTheme'
+import Box from '@mui/material/Box'
+
+const GurbIdentificationMember = (props) => {
+
+  const { values, setFieldValue } = props
+  const { t } = useTranslation()
+
+
+  return (
+    <Box sx={{ display: 'flex', flexDirection: 'column', gap: 1 }}>
+      <Typography sx={{ ...textHeader2, mb: 5 }}>{t('GURB_IDENTIFICATION')}</Typography>
+
+      TODO Identification Member
+    </Box>
+  )
+}
+
+export default GurbIdentificationMember
